Updates to Election-Contest-Audit relationship.
Per call this afternoon, the relationship between Election, Contest, and
Audit objects has been updated. Elections contain information to qualify
an entire election in general and a list of contests within the
Election. Contests contain all information about a specific contest.
Docstrings have been updated. Some imports have been modified. Election
and Contest classes have been separated. BRLA has been updated slightly
to inlcude another example method.
